refactor : [문자열 계산기 | 로또] ExpressionConverter -> ExpressionUtils, SplitString -> SplitStringUtils로 이름 변경
- 패키지 구조 정렬
