+++ src/main/java/calculator/domain/StringCalculator.java
@@ -1,4 +1,7 @@
-package calculator;
+package calculator.domain;
+
+import calculator.util.ExpressionUtils;
+import calculator.util.SplitStringUtils;
 
 import java.util.Deque;
 
@@ -9,7 +12,7 @@ public class StringCalculator {
     }
 
     public static int stringCalculate(String text) {
-        Deque<String> postFixExpressionStack = ExpressionConverter.convertToPostFixExpression(SplitString.split(text));
+        Deque<String> postFixExpressionStack = ExpressionUtils.convertToPostFixExpression(SplitStringUtils.split(text));
         while (!postFixExpressionStack.isEmpty()) {
             if (postFixExpressionStack.size() == ONLY_ONE_OPERAND) {
                 return new Operand(postFixExpressionStack.pollLast()).getNumber();
