+++ src/main/java/calculator/util/SplitStringUtils.java
@@ -1,20 +1,18 @@
-package lotto.util;
+package calculator.util;
 
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-public class SplitString {
-    private static final String COMMA_DELIMITER = ",";
+public class SplitStringUtils {
+    private static final String BLANK_DELIMITER = " ";
 
-    private SplitString() {
+    private SplitStringUtils() {
     }
 
     public static List<String> split(String text) {
         if (isBlank(text)) {
             throw new IllegalArgumentException("입력 문자열은 비어있을 수 없습니다.");
         }
-        return split(text, COMMA_DELIMITER);
+        return split(text, BLANK_DELIMITER);
     }
 
     private static boolean isBlank(String text) {
@@ -22,6 +20,6 @@ public class SplitString {
     }
 
     private static List<String> split(String text, String delimiter) {
-        return Stream.of(text.split(delimiter)).map(String::trim).collect(Collectors.toList());
+        return List.of(text.split(delimiter));
     }
 }
