+++ src/test/java/calculator/domain/StringCalculatorTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
