+++ src/test/java/calculator/util/SplitStringUtilsTest.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package calculator.util;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -9,15 +9,15 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-class SplitStringTest {
+class SplitStringUtilsTest {
     @Test
-    void split_문자열_콤마_구분() {
-        assertThat(SplitString.split(" 1,2,3 , 4,5 ,6 ")).isEqualTo(List.of("1", "2", "3", "4", "5", "6"));
+    void split_공백으로_구분() {
+        assertThat(SplitStringUtils.split("1 + 3 - 2 * 5 / 2")).isEqualTo(List.of("1", "+", "3", "-", "2", "*", "5", "/", "2"));
     }
 
     @ParameterizedTest
     @NullAndEmptySource
     void split_입력문자열_비어있는_경우(String elem) {
-        assertThatThrownBy(() -> SplitString.split(elem)).isInstanceOf(IllegalArgumentException.class);
+        assertThatThrownBy(() -> SplitStringUtils.split(elem)).isInstanceOf(IllegalArgumentException.class);
     }
 }
